% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_daq_to_hdf5.R
\name{read_daq_and_save_to_hdf5}
\alias{read_daq_and_save_to_hdf5}
\title{Convert a DAQ file to HDF4 using undaqTools package in Python}
\usage{
read_daq_and_save_to_hdf5(file_name)
}
\arguments{
\item{file_name}{path to the daq file (string)}
}
\value{
saves the converted hdf5 file to disk with the same name as the original daq file
}
\description{
Convert a DAQ file to HDF4 using undaqTools package in Python
}
\examples{
\dontrun{
library(reticulate)
use_virtualenv("r-reticulate")
read_daq_and_save_to_hdf5("path/to/daq/file/myfile.daq")
}
}
