% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_driver_data.R
\name{read_driver_data}
\alias{read_driver_data}
\title{Get several driving variables for the driver}
\usage{
read_driver_data(file_name, metric = TRUE)
}
\arguments{
\item{file_name}{path to daq or hdf5 file (string)}

\item{metric}{whether to get data in metric or english units. Default is TRUE (boolean)}
}
\value{
a data.frame containing several driving variables
}
\description{
Get several driving variables for the driver
}
\examples{
\dontrun{
library(reticulate)
use_virtualenv("r-reticulate")
read_driver_data("path/to/daq/file/myfile.daq")
read_driver_data("path/to/daq/file/myfile.hdf5")
}
}
